Add workaround for small eraser touches without pen events
If an eraser arrives without a pen event, and the eraser is held at an angle (which
makes it pen-sized), then the touch will be marked as eraser (and not pen) if the most
recent switch event was the eraser tip, until the switch dictates otherwise.  This fix
was in prior builds but got broken due to another fix.