Fixes for Pencil firmware update
1. Fixes UI related bugs where the update dialog sits at 100% indefinitely.
2. Provides a more robust TI-OAD service client implementation.

*Testing Done*

Many updates from 55 (factory) to 74. On iOS 7 always works. Even did
a test with the pencil across the room. It took 20 minutes but the update
succeeded.

On iOS 8.2 about 1/3 of the updates failed. The symptom is that right as
the last few blocks are being written the device simply resets. The UI
does the right thing with this change in messaging the failure to the
user.

Did a test where I walked away with the pencil after starting the update.
The UI messaged a failure properly.

Did many tests where I pulled and put back in the battery. The update
process has to start over from scratch but it does work.