Fixing a bug where the animations stop working in the pairing spot.
This happens because when the bluetooth radio is turned off the
"reset" method is called on the FTPenManager instance but the
pairing spot UI is shared statically with many other view instances.
Reset was unregistering the CADisplayLink used to do animation
and the animations would not occur until/unless the static factory
method was invoked again.

This change removes the reset of the CADisplayLink from reset and moves
it, instead, into the manager's shutdown method.

Testing done:
Manual testing with SDK app. Built and tested paper manually.