Fixing moved touch without begin assert (bugzid:5163)
As stated in the comments we want to revisit this issue and
have an SDK backlog item (https://www.pivotaltracker.com/story/show/102706136)
to do so. Testing on iPhone and iPad don't reveal any defects
caused by the multiple UITouch events. There is the potential
for some misclassifications and missed gestures but these would
only occur for edge cases and I was not able to specifically
identify any in my testing.